final_
ddd

+++ AdminApp/functions/images/adaugare_categorie_upload.php
@@ -22,16 +22,15 @@ if(isset($_FILES["file"]["type"]))
         {
             if (file_exists("upload/category" . $name)) {
                 echo $name . " <span id='invalid'><b>already exists.</b></span> ";
+                if(!$name=="default-category.png"){
                 unlink("upload/category".$name);
-                echo "<span id='success'>Image Deleted Successfully...!!</span><br/>";
+                }
+               // echo "<span id='success'>Image Deleted Successfully...!!</span><br/>";
                 $sourcePath = $_FILES['file']['tmp_name']; // Storing source path of the file in a variable
                 $targetPath = "upload/category".$name; // Target path where file is to be stored
                 move_uploaded_file($sourcePath,$targetPath) ; // Moving Uploaded file
-                echo "<span id='success'>Image Uploaded Successfully...!!</span><br/>";
-                echo "<br/><b>File Name:</b> " . $name . "<br>";
-                echo "<b>Type:</b> " . $_FILES["file"]["type"] . "<br>";
-                echo "<b>Size:</b> " . ($_FILES["file"]["size"] / 1024) . " kB<br>";
-                echo "<b>Temp file:</b> " . $_FILES["file"]["tmp_name"] . "<br>";
+                echo "Datele au fost actualizate cu succes!";
+              
                 //updateCategory($_POST['valoare-input'],$name,$_POST['id-cat'],$conn);
                 
                 adaugare_cat($_POST['valoare-input'], $name,$conn);
@@ -41,11 +40,8 @@ if(isset($_FILES["file"]["type"]))
                 $sourcePath = $_FILES['file']['tmp_name']; // Storing source path of the file in a variable
                 $targetPath = "upload/category/".$name; // Target path where file is to be stored
                 move_uploaded_file($sourcePath,$targetPath) ; // Moving Uploaded file
-                echo "<span id='success'>Image Uploaded Successfully...!!</span><br/>";
-                echo "<br/><b>File Name:</b> " . $name . "<br>";
-                echo "<b>Type:</b> " . $_FILES["file"]["type"] . "<br>";
-                echo "<b>Size:</b> " . ($_FILES["file"]["size"] / 1024) . " kB<br>";
-                echo "<b>Temp file:</b> " . $_FILES["file"]["tmp_name"] . "<br>";
+               // echo "Descrierea categoriei și imaginea au fost create cu succes!";
+                
                // updateCategory($_POST['valoare-input'],$name,$_POST['id-cat'],$conn);
                 
                 adaugare_cat($_POST['valoare-input'], $name,$conn);
@@ -54,7 +50,9 @@ if(isset($_FILES["file"]["type"]))
     }
     else
     {
-        echo "<span id='invalid'>***Invalid file Size or Type***<span>";
+       // echo "Categorie noua cu imagine default";    
+                
+        adaugare_cat($_POST['valoare-input'], "default-category.png",$conn);
     }
 }
 ?>
